start to separate out summaries to clean them
