+++ baus.py
@@ -9,6 +9,8 @@ from baus import slr
 from baus import earthquake
 from baus import ual
 from baus import validation
+from baus import geographic_summaries
+from baus import travel_model_summaries
 import numpy as np
 import pandas as pd
 import orca
@@ -359,7 +361,6 @@ def get_baseyear_models():
         "hazards_slr_summary",
         "hazards_eq_summary",
         "diagnostic_output",
-        "environment_config",
         "slack_report"
     ]
